src/MB.Backend/pyproject.toml
[tool.poetry]
name = "MB.Backend"
version = "0.1.0"
description = ""
authors = ["Jedrzej Szadejko <jendziu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
